Transfered the ranking endpoint
