+++ src/data_api.py
@@ -1,7 +1,3 @@
-# flake8: noqa: E402
-#
-# We need to configure logging as the first thing to do...
-#
 import logging
 import traceback
 
@@ -19,8 +15,6 @@ from datetime import datetime as datetime
 from datetime import timedelta
 from urllib.parse import urljoin, urlparse, urlunparse
 
-import pandas as pd
-import psycopg2
 import redis
 import requests
 import yaml
@@ -37,7 +31,6 @@ import execution
 
 # Input schemata for validating several API requests
 import schema
-import sql_utils
 
 # Auxiliary custom functions & SQL query templates for ranking
 import utils
@@ -56,11 +49,8 @@ from routes.settings import settings_bp
 from routes.users import api_user_editor, users_bp
 from routes.workflows import workflows_bp
 from routes.llm_search import llmsearch_bp
+from routes.ranking import ranking_bp
 
-# fmt: on
-
-
-# from container_utils import create_container
 
 logger = logging.getLogger(__name__)
 
@@ -103,6 +93,7 @@ app.register_blueprint(auth_tool_bp, url_prefix="/api/v1/auth")
 app.register_blueprint(catalog_bp, url_prefix="/api/v2")
 app.register_blueprint(workflows_bp, url_prefix="/api/v2")
 app.register_blueprint(llmsearch_bp, url_prefix="/api/v2/llm")
+app.register_blueprint(ranking_bp, url_prefix="/api/v1")
 
 logger.info("Blueprints registered successfully.")
 logger.debug("Endpoints: %s", app.url_map)
@@ -635,292 +626,6 @@ def api_task_parameters(query_data):
     )
 
     return jsonify(json.loads(response.text))
-
-
-################################## RANKING OPERATIONS ########################################
-@app.route("/api/v1/catalog/rank", methods=["POST"])
-@app.input(
-    schema.Ranking,
-    location="json",
-    example={
-        "rank_preferences": {
-            "tags": ["Geospatial", "POI"],
-            "theme": ["Land Use", "Land Cover", "Imagery"],
-            "language": ["en", "el", "fr"],
-            "spatial": {
-                "type": "Polygon",
-                "coordinates": [
-                    [
-                        [12.362, 45.39],
-                        [12.485, 45.39],
-                        [12.485, 45.576],
-                        [12.362, 45.576],
-                        [12.362, 45.39],
-                    ]
-                ],
-            },
-        },
-        "settings": {"k": 10, "algorithm": "threshold", "weights": [0.3, 0.5, 0.4]},
-    },
-)
-@app.output(schema.ResponseOK, status_code=200)
-@app.doc(tags=["Ranking Operations"], security=security_doc)
-@token_active
-def api_catalog_rank(json_data):
-    """Submit a rank request regarding specific metadata attributes (facets) to the Data Catalog.
-
-    Args:
-        json_data: A JSON with facet preferences for searching in the Data Catalog. Facet name should match a property specified in the STELAR Ontology.
-
-    Returns:
-        A JSON with datasets ranked by the specified facet(s). The matching score per facet criterion is also listed per returned dataset.
-    """
-
-    # EXAMPLE: curl -X POST -H 'Content-Type: application/json' http://127.0.0.1:9055/api/v1/catalog/rank -d '{"q":{"theme":"POI"}}'
-
-    config = current_app.config["settings"]
-
-    if request.headers:
-        if request.headers.get("Api-Token") is not None:
-            package_headers, resource_headers = utils.create_CKAN_headers(
-                request.headers["Api-Token"]
-            )
-        else:
-            response = {
-                "success": False,
-                "help": request.url,
-                "error": {
-                    "__type": "Authorization Error",
-                    "name": [
-                        "No API_TOKEN specified. Please specify a valid API_TOKEN in the headers of your request."
-                    ],
-                },
-            }
-            return jsonify(response)
-    else:
-        response = {
-            "success": False,
-            "help": request.url,
-            "error": {
-                "__type": "Authorization Error",
-                "name": [
-                    "No headers specified. Please specify headers for your request, including a valid API TOKEN."
-                ],
-            },
-        }
-        return jsonify(response)
-
-    sql = ""
-    sql_id_filter = ""
-    ids = []
-    dict_df_facet_scores = (
-        {}
-    )  # dictionary with the returned input ranked lists per facet (key -> dataframe)
-    k = config["RANK_MAX_TOPK"]  # default top-k value (if not user-specified)
-    if request.data:
-        specs = json.loads(request.data.decode("utf-8"))
-        actual_profile_attributes = set(specs["filter_preferences"].keys()).union(
-            set(specs["rank_preferences"].keys())
-        )
-        #        print("INITIAL PROFILE ATTRIBUTES", actual_profile_attributes)
-
-        # STAGE #1: text-based keyword search targets SOLR (search engine for CKAN)
-        if "keywords" in specs:  # CASE #1(a): new keyword search
-            q = "?q=" + ",".join("'{0}'".format(kw) for kw in specs["keywords"])
-            #            print(q)
-            # Submit a preliminary search request to CKAN to find packages qualifying to the specified keywords
-            # Also include private datasets of the user's organization in the results
-            resp_basic = requests.get(
-                config["CKAN_API"]
-                + "package_search"
-                + q
-                + "&include_private=True&fl=*,score&rows="
-                + str(config["RANK_MAX_TOPK"])
-                + "&start=0",
-                headers=package_headers,
-            )
-            if resp_basic.status_code == 200:
-                json_resp_basic = resp_basic.json()
-                # FIXME: Handle large number of returned id's -> not efficient when filtering with SQL
-                if json_resp_basic["success"]:  # Results from keyword-based search only
-                    results = json_resp_basic["result"]["results"]
-                    ids = [res["id"] for res in results if "id" in res]
-        #                    print('keyword results:',len(ids))
-        elif (
-            "ids" in specs
-        ):  # CASE #1(b): Identifiers of datasets already qualifying keyword search criteria
-            if len(specs["ids"]) > 0:
-                ids = specs["ids"]
-        if len(ids) > 0:  # Specify the previously filtered items to be sent for ranking
-            sql_id_filter, k = utils.format_sql_filter(ids)
-        else:  # No results from filtering, no sense to continue with further filtering
-            response = {
-                "help": request.url,
-                "result": {
-                    "count": 0,
-                    "facets": {},
-                    "results": [],
-                    "sort": "score desc, metadata_modified desc",
-                },
-                "success": True,
-            }
-            return jsonify(response)
-
-        # STAGE #2: Apply any filtering criteria (NOT participating in the ranking)
-        if "filter_preferences" in specs:
-            filter_sql_commands = utils.format_facet_preferences(
-                specs["filter_preferences"], sql_id_filter, config["RANK_MAX_TOPK"]
-            )
-            # Submit each SELECT query to the PostgreSQL database with the respective parameters
-            # IMPORTANT! PostgreSQL credentials are required to complete this request
-            for key in filter_sql_commands.keys():
-                sql = filter_sql_commands[key]
-                results = utils.execSql(sql)
-                #                print(len(results), sql)
-                filter_ids = [res["id"] for res in results if "id" in res]
-                if sql_id_filter == "":  # No keywords specified in search bar
-                    ids = filter_ids
-                else:  # Keep only matching id's
-                    ids = [id for id in ids if id in filter_ids]
-        #                print(key, len(ids))
-        if len(ids) > 0:  # Specify the previously filtered items to be sent for ranking
-            sql_id_filter, k = utils.format_sql_filter(ids)
-        else:  # No results from filtering, no sense to apply ranking
-            response = {
-                "help": request.url,
-                "result": {
-                    "count": 0,
-                    "facets": {},
-                    "results": [],
-                    "sort": "score desc, metadata_modified desc",
-                },
-                "success": True,
-            }
-            return jsonify(response)
-
-        # STAGE #3: Prepare SQL queries for each of the ranking preferences
-        if "rank_preferences" in specs:
-            rank_sql_commands = utils.format_facet_preferences(
-                specs["rank_preferences"], sql_id_filter, config["RANK_MAX_TOPK"]
-            )
-            # FIXME: REMOVE IF HANDLED BY THE FRONT-END
-            # Examine settings for ranking
-            #            if 'settings' in specs:
-            #                if not 'algorithm' in specs['settings']: # Apply default algorithm for rank aggregation
-            #                    specs['settings']['algorithm'] = config['RANK_AGG_ALGORITHM']
-            #                if 'k' in specs['settings']:
-            #                    k = specs['settings']['k']
-            #                else:
-            #                    specs['settings']['k'] = k    # Rank aggregation library requires the total number of items
-            #                    print(specs['settings']['k'])
-            #            else:  # Specify default values for rank aggregation
-            #                specs['settings'] = {}
-            #                specs['settings']['algorithm'] = config['RANK_AGG_ALGORITHM']
-            #                specs['settings']['k'] = k
-
-            # Submit each SELECT query to the PostgreSQL database with the respective parameters
-            # IMPORTANT! PostgreSQL credentials are required to complete this request
-            input_lists = []
-            for key in rank_sql_commands.keys():
-                sql = rank_sql_commands[key]
-                #                print(key, '->', sql)
-                results = utils.execSql(sql)
-                #                print("FIELDS", key, len(results), len(ids))
-                # Fill any missing scores in the partial list for this facet
-                for id in ids:
-                    if id not in [d["id"] for d in results if "id" in d]:
-                        results.append({"id": id, "score": 0.0})
-                dict_df_facet_scores[key] = utils.read_list_json(results)
-            #                # In case a 'value' column (concerning PROFILING) is returned in results, remember to include its values in the final results
-            #                if 'value' in dict_df_facet_scores[key].columns:
-            #                    profile_attributes.append(key)
-            #                    print(key)
-
-            # Fetch values for all profiling metadata elements by submitting a SELECT query to the PostgreSQL database for the collected ids
-            # IMPORTANT! PostgreSQL credentials are required to complete this request
-            actual_profile_attributes = actual_profile_attributes.intersection(
-                utils.profile_attributes
-            )
-            #            print("ACTUAL PROFILE ATTRIBUTES", actual_profile_attributes)
-            for key in list(
-                actual_profile_attributes
-            ):  # list(set(utils.profile_attributes) - set(rank_sql_commands.keys())):
-                sql = utils.identifiers_sql_filter_template.replace(
-                    "_VIEW", utils.sql_views[key]
-                ).replace("_IDS", sql_id_filter)
-                #                print(key, '->', sql)
-                results = utils.execSql(sql)
-                print("PROFILING", key, len(results), len(ids))
-                # Fill any missing scores in the partial list for this facet
-                for id in ids:
-                    if id not in [d["id"] for d in results if "id" in d]:
-                        results.append({"id": id, "score": 0.0})
-                dict_df_facet_scores[key] = utils.read_list_json(results)
-
-                input_lists.append(dict_df_facet_scores[key])
-
-            # FIXME: REMOVE IF HANDLED BY THE FRONT-END
-            agg_scores = (
-                pd.DataFrame()
-            )  # No aggregated scores, report the original SOLR scores
-
-            # Compute the final ranked list of all items applying the specified rank aggregation method (e.g., threshold)
-        #            agg_scores = ranking.combined_ranking(input_lists, specs['settings'])
-        #            ids = agg_scores.index.values  # In case no keywords and no filter criteria have been spcified; only rank preferences
-        #            print(agg_scores.index.values)
-        elif (
-            "settings" in specs
-        ):  # Settings for rank aggregation assume at least once facet specification
-            response = {
-                "success": False,
-                "help": request.url,
-                "error": {
-                    "__type": "Incorrect specifications",
-                    "name": [
-                        "Incorrect or no facet preferences provided to rank items in the Data Catalog. Please specify at least one facet preference in a dictionary."
-                    ],
-                },
-            }
-            return jsonify(response)
-        else:  # No ranking to be applied; only search filters
-            specs["rank_preferences"] = {}  # Facets for ranking not specified
-            agg_scores = (
-                pd.DataFrame()
-            )  # No aggregated scores, report the original SOLR scores
-    else:
-        response = {
-            "success": False,
-            "help": request.url,
-            "error": {
-                "__type": "No specifications",
-                "name": [
-                    "No facet preferences provided to rank items in the Data Catalog. Please specify at least one facet preference in a dictionary."
-                ],
-            },
-        }
-        return jsonify(response)
-
-    # Retrieve from CKAN all metadata for the datasets in the final (aggregated ranked) list
-    # Also include private datasets of the user's organization in the results
-    q = "?q=" + " OR ".join(["id:" + id for id in ids])
-    response = requests.get(
-        config["CKAN_API"]
-        + "package_search"
-        + q
-        + "&rows="
-        + str(config["RANK_MAX_TOPK"])
-        + "&start=0&include_private=True",
-        headers=package_headers,
-    )
-
-    # Return the final list of results (the top-k ranked ones in case that ranking preferences are specified)
-    return utils.assign_scores(
-        response,
-        agg_scores,
-        dict_df_facet_scores,
-        specs["rank_preferences"],
-        list(actual_profile_attributes),
-    )
 
 
 ############################### PUBLISHING OPERATIONS ############################
